+++ src/components/containers/MainContainer.jsx
@@ -1,27 +1,12 @@
 import React, { Component } from 'react';
-import RoutesContainer from './RoutesContainer';
+import MainView from '../views/MainView';
 
 // Container component;
 class MainContainer extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      modalDisplay: false
-    }
-  }
-
-  //following two functions for opening/closing modal on local state
-  onOpenModal = () => {
-    this.setState({ modalDisplay: true });
-  };
-
-  onCloseModal = () => {
-    this.setState({ modalDisplay: false });
-  };
 
   render() {
     return (
-      <RoutesContainer />
+      <MainView />
     );
   }
 }
