+++ src/components/containers/index.js
@@ -3,6 +3,6 @@ export { default as HeaderContainer } from './HeaderContainer';
 export { default as MainContainer } from './MainContainer';
 export { default as FooterContainer } from './FooterContainer';
 export { default as RoutesContainer } from './RoutesContainer';
-export { default as LandingContainer } from './LandingContainer';
+export { default as LandingContainer } from './modules/Main/Landing/LandingContainer';
 export { default as LoginFormContainer } from './LoginFormContainer';
 export { default as LogoutContainer } from './LogoutContainer';
