+++ src/components/views/MainView.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import RoutesContainer from '../containers/RoutesContainer';
+
+const MainView = props => {
+  return (
+    <div className="section-main">
+      <RoutesContainer />
+    </div>
+  )
+};
+
+export default MainView;
