+++ src/components/views/index.js
@@ -1,6 +1,6 @@
 // Barrel file for presentational components, which will be used to compose our smart containers
 export { default as HeaderView } from './HeaderView';
-export { default as LandingView } from './LandingView';
+export { default as LandingView } from './modules/Main/Landing/LandingView';
 export { default as FooterView } from './FooterView';
-export { default as LoginFormView } from './LoginFormView';
-export { default as LogoutView } from './LogoutView';
+export { default as LoginFormView } from './modules/Header/LoginFormView';
+export { default as LogoutView } from './modules/Header/LogoutView';
