+++ src/components/views/modules/index.js
@@ -1,4 +1,4 @@
 // Barrel file for presentational components, which will be used to compose our smart containers
 export { default as AllBarbersView } from './AllBarbersView';
-export { default as AppointmentFormView } from './AppointmentFormView';
+export { default as AppointmentFormView } from './Main/Landing/AppointmentFormView';
 export { default as BarberAppointmentsView } from './BarberAppointmentsView';
